+++ changelog.md
@@ -2,6 +2,7 @@
 
 ## unreleased
 * Show error when input is higher than allowed
+* Allow selecting only a few package IDs
 
 ## 1.1.4 (2019-11-18)
 * Change date format for booking date field in contact forms